// src/api/model/loadResult.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * CRM.Backend.Web | v1
 * OpenAPI spec version: 1.0.0
 */

export interface LoadResult {
	/** @nullable */
	data?: unknown[] | null;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	totalCount?: number | string;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	groupCount?: number | string;
	/** @nullable */
	summary?: unknown[] | null;
}