// src/api/model/projectDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * CRM.Backend.Web | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ProjectDto {
	id?: string;
	/** @nullable */
	name?: string | null;
	/** @nullable */
	imgUrl?: string | null;
	/** @nullable */
	logoUrl?: string | null;
	/** @nullable */
	description?: string | null;
	/** @nullable */
	fileName?: string | null;
	/** @nullable */
	fileType?: string | null;
	/** @nullable */
	fileUrl?: string | null;
	workspaceId?: string;
}