// src/api/model/workspaceUserDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * CRM.Backend.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { DateFormatOption } from "./dateFormatOption";
import type { InvitationStatus } from "./invitationStatus";
import type { LanguageOption } from "./languageOption";
import type { ThemeOption } from "./themeOption";
import type { TimeFormatOption } from "./timeFormatOption";
import type { TimezoneOption } from "./timezoneOption";

export interface WorkspaceUserDto {
	id?: string;
	workspaceId?: string;
	/** @nullable */
	userId?: string | null;
	/** @nullable */
	customRoleId?: string | null;
	/** @nullable */
	customRoleName?: string | null;
	/** @nullable */
	joinedAt?: string | null;
	isActive?: boolean;
	/** @nullable */
	email?: string | null;
	/** @nullable */
	invitationToken?: string | null;
	/** @nullable */
	invitationExpiresAt?: string | null;
	invitationStatus?: InvitationStatus;
	notificationsEnabled?: boolean;
	emailNotifications?: boolean;
	language?: LanguageOption;
	theme?: ThemeOption;
	timezone?: TimezoneOption;
	timeFormat?: TimeFormatOption;
	dateFormat?: DateFormatOption;
}